tests/wpt/web-platform-tests/workers/constructors/SharedWorker/unresolvable-url.html

<!--
/*
-->
<!doctype html>
<title>resolving broken url</title>
<link rel=help href="http://www.whatwg.org/html/#dom-sharedworker">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws("SyntaxError", function() {
    var worker = new SharedWorker('http://foo bar');
  });
});
</script>
<!--
*/
//-->
